WIP: resource graph
