+++ packages/konsole-app/src/components/RegistryGraph.tsx
@@ -6,26 +6,34 @@ import { Box } from '@mui/material';
 import React, { useEffect, useState } from 'react';
 import useResizeAware from 'react-resize-aware';
 
-import { SVG, convertTreeToGraph, createTree, useGrid } from '@dxos/gem-core';
+import { SVG, convertTreeToGraph, createTree, useGrid, GraphType } from '@dxos/gem-core';
 import { ForceLayout, Graph } from '@dxos/gem-spore';
-import { Domain } from '@dxos/registry-client';
+import { Domain, Resource } from '@dxos/registry-client';
 
 import { IRecord } from './RecordsTable';
 
-interface RecordsGraphProps {
+interface RegistryGraphProps {
   domains?: Domain[]
-  records?: IRecord[]
+  records?: IRecord[],
+  resources?: Resource[]
 }
 
 // https://github.com/dxos/gem/blob/main/packages/gem-spore/stories/graph.stories.tsx
-export const RecordsGraph = ({ domains = [], records = [] }: RecordsGraphProps) => {
+export const RegistryGraph = ({ domains = [], records = [], resources = [] }: RegistryGraphProps) => {
   const [resizeListener, size] = useResizeAware();
   const grid = useGrid(size);
-  const [data] = useState(convertTreeToGraph(createTree({ minDepth: 1, maxDepth: 3 })));
+  const [data, setData] = useState<GraphType>({links: [], nodes: []});
   const [layout] = useState(() => new ForceLayout({ force: { links: { distance: 80 } } }));
+
+  console.log({domains, records, resources, data})
+
   useEffect(() => {
-    // TODO(burdon): Generate graph.
-  }, [domains, records]);
+    const resourceNodes: GraphType['nodes'] = resources.map(resource => ({id: resource.id.toString(), title: resource.id.toString()}))
+    const recordNodes: GraphType['nodes'] = records.map(record => ({id: record.cid.toString(), title: record.description ?? record.cid.toString()}))
+    const nodes = [...resourceNodes, ...recordNodes]
+    const links: GraphType['links'] = []
+    setData({nodes, links})
+  }, [domains, records, resources]);
 
   return (
     <Box
