+++ packages/konsole-app/src/components/index.ts
@@ -9,7 +9,6 @@ export * from './Fullscreen';
 export * from './JsonView';
 export * from './LogTable';
 export * from './Panel';
-export * from './RecordsGraph';
 export * from './RecordsTable';
 export * from './RecordTypeSelector';
 export * from './ResourceRecordsTable';
