+++ packages/konsole-app/src/containers/RecordsPanel.tsx
@@ -18,7 +18,6 @@ import {
   IRecord,
   JsonView,
   Panel,
-  RecordsGraph,
   RecordsTable,
   RecordTypeSelector,
   SearchBar,
@@ -74,46 +73,6 @@ export const joinRecords = (records: RegistryRecord[], recordTypes: RegistryType
   });
 };
 
-const views = [
-  { key: 'table', Icon: TableIcon },
-  { key: 'graph', Icon: GraphIcon }
-];
-
-const ViewSelector = ({ view, onChange }: { view: string, onChange: (view: string) => void }) => {
-  const handleChange = (event: React.MouseEvent<HTMLElement>, view: string) => {
-    onChange(view);
-  };
-
-  return (
-    <ToggleButtonGroup
-      exclusive
-      value={view}
-      onChange={handleChange}
-      size='small'
-      sx={{
-        marginRight: 4
-      }}
-    >
-      {views.map(({ key, Icon }) => (
-        <ToggleButton
-          key={key}
-          value={key}
-          color='primary'
-        >
-          <Icon />
-        </ToggleButton>
-      ))}
-    </ToggleButtonGroup>
-  );
-};
-
-// TODO(burdon): Render is not visible (to maintain state).
-const ViewPanel = ({ children, visible }: { children: React.ReactNode, visible: boolean }) => (
-  <>
-    {visible && children}
-  </>
-);
-
 /**
  * Display records panel
  */
@@ -123,12 +82,10 @@ export const RecordsPanel = ({ match }: { match?: any }) => {
   const { cid }: { cid?: string } = useParams();
   const selected = safe<CID | undefined>(() => cid ? CID.fromB58String(cid) : undefined);
 
-  const [view, setView] = useState(views[0].key);
   const [recordType, setRecordType] = useState<CID | undefined>(undefined);
   const [search, setSearch] = useState<string | undefined>();
   const query = useMemo<IQuery>(() => ({ type: recordType, text: search }), [recordType, search]);
 
-  const { domains } = useDomains();
   const { recordTypes } = useRecordTypes();
   const { records: registryRecords } = useRecords(query);
   const records = joinRecords(registryRecords, recordTypes, config);
@@ -149,7 +106,6 @@ export const RecordsPanel = ({ match }: { match?: any }) => {
     <Panel
       toolbar={(
         <Toolbar>
-          <ViewSelector view={view} onChange={setView} />
           <Box>
             <RecordTypeSelector
               types={recordTypes}
@@ -183,31 +139,23 @@ export const RecordsPanel = ({ match }: { match?: any }) => {
         </Toolbar>
       )}
     >
-      <ViewPanel visible={view === 'table'}>
-        <RecordsTable
-          records={records}
-          selected={selected}
-          onSelect={handleSelect}
-        />
-        <Collapse in={selected !== undefined} timeout='auto' unmountOnExit>
-          <Paper
-            sx={{
-              marginTop: 1,
-              height: 304,
-              overflow: 'scroll',
-              padding: 1
-            }}
-          >
-            <JsonView data={selected && records.find(record => record.cid.equals(selected.toB58String()))} />
-          </Paper>
-        </Collapse>
-      </ViewPanel>
-      <ViewPanel visible={view === 'graph'}>
-        <RecordsGraph
-          domains={domains}
-          records={records}
-        />
-      </ViewPanel>
+      <RecordsTable
+        records={records}
+        selected={selected}
+        onSelect={handleSelect}
+      />
+      <Collapse in={selected !== undefined} timeout='auto' unmountOnExit>
+        <Paper
+          sx={{
+            marginTop: 1,
+            height: 304,
+            overflow: 'scroll',
+            padding: 1
+          }}
+        >
+          <JsonView data={selected && records.find(record => record.cid.equals(selected.toB58String()))} />
+        </Paper>
+      </Collapse>
     </Panel>
   );
 };
